Fix deprecation.

+++ resources/mathpublisher.php
@@ -536,7 +536,7 @@ function affiche_symbol($texte, $haut)
 			$tmp_blanc = imagecolorallocate($tmp_img, 255, 255, 255);
 			$tmp_blanc = imagecolortransparent($tmp_img, $tmp_blanc);
 			imagefilledrectangle($tmp_img, 0, 0, $tmp_largeur, $tmp_hauteur, $tmp_blanc);
-			imagettftext($tmp_img, $t, 0, 5, $tmp_hauteur / 2, $tmp_noir, $font, $texte);
+			imagettftext($tmp_img, $t, 0, 5, (int)($tmp_hauteur / 2), $tmp_noir, $font, $texte);
 			$img = $tmp_img;
 			$toutblanc = true;
 			$sx = $sy = $ex = $ey = -1;
@@ -1061,9 +1061,9 @@ class expression_math extends expression
 		$blanc = imagecolortransparent($result, $blanc);
 		$this->base_verticale = $hauteur1 + 2;
 		imagefilledrectangle($result, 0, 0, $largeur + 4, $hauteur - 1, $blanc);
-		imagecopy($result, $img1, ($largeur - $largeur1) / 2, 0, 0, 0, $largeur1, $hauteur1);
+		imagecopy($result, $img1, (int)(($largeur - $largeur1) / 2), 0, 0, 0, $largeur1, $hauteur1);
 		imageline($result, 0, $this->base_verticale, $largeur, $this->base_verticale, $noir);
-		imagecopy($result, $img2, ($largeur - $largeur2) / 2, $hauteur1 + 4, 0, 0, $largeur2, $hauteur2);
+		imagecopy($result, $img2, (int)(($largeur - $largeur2) / 2), $hauteur1 + 4, 0, 0, $largeur2, $hauteur2);
 		$this->image = $result;
 	}
 
@@ -1081,24 +1081,24 @@ class expression_math extends expression
 		$largeur2 = imagesx($img2);
 		$largeur = $largeur1 + $largeur2;
 		if ($hauteur1 >= $hauteur2) {
-			$hauteur = ceil($hauteur2 / 2 + $hauteur1);
+			$hauteur = (int)ceil($hauteur2 / 2 + $hauteur1);
 			$this->base_verticale = $hauteur2 / 2 + $base1;
 			$result = imagecreate(max($largeur, 1), max($hauteur, 1));
 			$noir = imagecolorallocate($result, 0, 0, 0);
 			$blanc = imagecolorallocate($result, 255, 255, 255);
 			$blanc = imagecolortransparent($result, $blanc);
 			imagefilledrectangle($result, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-			imagecopy($result, $img1, 0, ceil($hauteur2 / 2), 0, 0, $largeur1, $hauteur1);
+			imagecopy($result, $img1, 0, (int)ceil($hauteur2 / 2), 0, 0, $largeur1, $hauteur1);
 			imagecopy($result, $img2, $largeur1, 0, 0, 0, $largeur2, $hauteur2);
 		} else {
-			$hauteur = ceil($hauteur1 / 2 + $hauteur2);
+			$hauteur = (int)ceil($hauteur1 / 2 + $hauteur2);
 			$this->base_verticale = $hauteur2 - $base1 + $hauteur1 / 2;
 			$result = imagecreate(max($largeur, 1), max($hauteur, 1));
 			$noir = imagecolorallocate($result, 0, 0, 0);
 			$blanc = imagecolorallocate($result, 255, 255, 255);
 			$blanc = imagecolortransparent($result, $blanc);
 			imagefilledrectangle($result, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-			imagecopy($result, $img1, 0, ceil($hauteur2 - $hauteur1 / 2), 0, 0, $largeur1, $hauteur1);
+			imagecopy($result, $img1, 0, (int)ceil($hauteur2 - $hauteur1 / 2), 0, 0, $largeur1, $hauteur1);
 			imagecopy($result, $img2, $largeur1, 0, 0, 0, $largeur2, $hauteur2);
 		}
 		$this->image = $result;
@@ -1118,7 +1118,7 @@ class expression_math extends expression
 		$largeur2 = imagesx($img2);
 		$largeur = $largeur1 + $largeur2;
 		if ($hauteur1 >= $hauteur2) {
-			$hauteur = ceil($hauteur2 / 2 + $hauteur1);
+			$hauteur = (int)ceil($hauteur2 / 2 + $hauteur1);
 			$this->base_verticale = $base1;
 			$result = imagecreate(max($largeur, 1), max($hauteur, 1));
 			$noir = imagecolorallocate($result, 0, 0, 0);
@@ -1126,9 +1126,9 @@ class expression_math extends expression
 			$blanc = imagecolortransparent($result, $blanc);
 			imagefilledrectangle($result, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
 			imagecopy($result, $img1, 0, 0, 0, 0, $largeur1, $hauteur1);
-			imagecopy($result, $img2, $largeur1, ceil($hauteur1 - $hauteur2 / 2), 0, 0, $largeur2, $hauteur2);
+			imagecopy($result, $img2, $largeur1, (int)ceil($hauteur1 - $hauteur2 / 2), 0, 0, $largeur2, $hauteur2);
 		} else {
-			$hauteur = ceil($hauteur1 / 2 + $hauteur2);
+			$hauteur = (int)ceil($hauteur1 / 2 + $hauteur2);
 			$this->base_verticale = $base1;
 			$result = imagecreate(max($largeur, 1), max($hauteur, 1));
 			$noir = imagecolorallocate($result, 0, 0, 0);
@@ -1136,7 +1136,7 @@ class expression_math extends expression
 			$blanc = imagecolortransparent($result, $blanc);
 			imagefilledrectangle($result, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
 			imagecopy($result, $img1, 0, 0, 0, 0, $largeur1, $hauteur1);
-			imagecopy($result, $img2, $largeur1, ceil($hauteur1 / 2), 0, 0, $largeur2, $hauteur2);
+			imagecopy($result, $img2, $largeur1, (int)ceil($hauteur1 / 2), 0, 0, $largeur2, $hauteur2);
 		}
 		$this->image = $result;
 	}
@@ -1237,9 +1237,9 @@ class expression_math extends expression
 		$blanc = imagecolorallocate($imggauche, 255, 255, 255);
 		$blanc = imagecolortransparent($imggauche, $blanc);
 		imagefilledrectangle($imggauche, 0, 0, $largeurgauche - 1, $hauteurgauche - 1, $blanc);
-		imagecopy($imggauche, $imgsymbole, ($largeurgauche - $largeursymbole) / 2, $hauteur2, 0, 0, $largeursymbole, $hauteursymbole);
-		imagecopy($imggauche, $img2, ($largeurgauche - $largeur2) / 2, 0, 0, 0, $largeur2, $hauteur2);
-		imagecopy($imggauche, $img1, ($largeurgauche - $largeur1) / 2, $hauteur2 + $hauteursymbole, 0, 0, $largeur1, $hauteur1);
+		imagecopy($imggauche, $imgsymbole, (int)(($largeurgauche - $largeursymbole) / 2), $hauteur2, 0, 0, $largeursymbole, $hauteursymbole);
+		imagecopy($imggauche, $img2, (int)(($largeurgauche - $largeur2) / 2), 0, 0, 0, $largeur2, $hauteur2);
+		imagecopy($imggauche, $img1, (int)(($largeurgauche - $largeur1) / 2), $hauteur2 + $hauteursymbole, 0, 0, $largeur1, $hauteur1);
 		$imgfin = alignement2($imggauche, $basesymbole + $hauteur2, $imgexp, $baseexp);
 		$this->image = $imgfin;
 		$this->base_verticale = max($basesymbole + $hauteur2, $baseexp + $hauteur2);
@@ -1267,8 +1267,8 @@ class expression_math extends expression
 		$blanc = imagecolorallocate($imgfin, 255, 255, 255);
 		$blanc = imagecolortransparent($imgfin, $blanc);
 		imagefilledrectangle($imgfin, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-		imagecopy($imgfin, $imgsup, ($largeur - $largeursup) / 2, 0, 0, 0, $largeursup, $hauteursup);
-		imagecopy($imgfin, $imgexp, ($largeur - $largeurexp) / 2, $hauteursup, 0, 0, $largeurexp, $hauteurexp);
+		imagecopy($imgfin, $imgsup, (int)(($largeur - $largeursup) / 2), 0, 0, 0, $largeursup, $hauteursup);
+		imagecopy($imgfin, $imgexp, (int)(($largeur - $largeurexp) / 2), $hauteursup, 0, 0, $largeurexp, $hauteurexp);
 		$this->image = $imgfin;
 		$this->base_verticale = $baseexp + $hauteursup;
 	}
@@ -1295,8 +1295,8 @@ class expression_math extends expression
 		$blanc = imagecolorallocate($imgfin, 255, 255, 255);
 		$blanc = imagecolortransparent($imgfin, $blanc);
 		imagefilledrectangle($imgfin, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-		imagecopy($imgfin, $imgexp, ($largeur - $largeurexp) / 2, 0, 0, 0, $largeurexp, $hauteurexp);
-		imagecopy($imgfin, $imginf, ($largeur - $largeurinf) / 2, $hauteurexp, 0, 0, $largeurinf, $hauteurinf);
+		imagecopy($imgfin, $imgexp, (int)(($largeur - $largeurexp) / 2), 0, 0, 0, $largeurexp, $hauteurexp);
+		imagecopy($imgfin, $imginf, (int)(($largeur - $largeurinf) / 2), $hauteurexp, 0, 0, $largeurinf, $hauteurinf);
 		$this->image = $imgfin;
 		$this->base_verticale = $baseexp;
 	}
@@ -1354,7 +1354,7 @@ class expression_math extends expression
 			$l = $padding / 2 - 1;
 			for ($col = 0; $col < $nbcolonne; $col++) {
 				if ($i < count($this->noeuds[3]->noeuds)) {
-					imagecopy($imgfin, $img[$i], $l + ceil($largeur_colonne[$col] - $largeur[$i]) / 2, $h + $dessus_ligne[$ligne] - $base[$i], 0, 0, $largeur[$i], $hauteur[$i]);
+					imagecopy($imgfin, $img[$i], (int)($l + ceil($largeur_colonne[$col] - $largeur[$i]) / 2), $h + $dessus_ligne[$ligne] - $base[$i], 0, 0, $largeur[$i], $hauteur[$i]);
 					//ImageRectangle($imgfin,$l,$h,$l+$largeur_colonne[$col],$h+$hauteur_ligne[$ligne],$noir);
 				}
 				$l+=$largeur_colonne[$col] + $padding;
@@ -1422,20 +1422,20 @@ class expression_math extends expression
 		for ($ligne = 0; $ligne < $nbligne; $ligne++) {
 			$l = $padding / 2 - 1;
 			if (substr($typecolonne, 0, 1) == "1") {
-				ImageLine($imgfin, 0, $h - $padding / 2, 0, $h + $hauteur_ligne[$ligne] + $padding / 2, $noir);
+				ImageLine($imgfin, 0, (int)($h - $padding / 2), 0, (int)($h + $hauteur_ligne[$ligne] + $padding / 2), $noir);
 			}
 			for ($col = 0; $col < $nbcolonne; $col++) {
 				if ($i < count($this->noeuds[3]->noeuds)) {
-					imagecopy($imgfin, $img[$i], $l + ceil($largeur_colonne[$col] - $largeur[$i]) / 2, $h + $dessus_ligne[$ligne] - $base[$i], 0, 0, $largeur[$i], $hauteur[$i]);
+					imagecopy($imgfin, $img[$i], (int)($l + ceil($largeur_colonne[$col] - $largeur[$i]) / 2), $h + $dessus_ligne[$ligne] - $base[$i], 0, 0, $largeur[$i], $hauteur[$i]);
 					if (substr($typecolonne, $col + 1, 1) == "1") {
-						ImageLine($imgfin, $l + $largeur_colonne[$col] + $padding / 2, $h - $padding / 2, $l + $largeur_colonne[$col] + $padding / 2, $h + $hauteur_ligne[$ligne] + $padding / 2, $noir);
+						ImageLine($imgfin, (int)($l + $largeur_colonne[$col] + $padding / 2), (int)($h - $padding / 2), (int)($l + $largeur_colonne[$col] + $padding / 2), (int)($h + $hauteur_ligne[$ligne] + $padding / 2), $noir);
 					}
 				}
 				$l+=$largeur_colonne[$col] + $padding;
 				$i++;
 			}
 			if (substr($typeligne, $ligne + 1, 1) == "1") {
-				ImageLine($imgfin, 0, $h + $hauteur_ligne[$ligne] + $padding / 2, $largeurfin - 1, $h + $hauteur_ligne[$ligne] + $padding / 2, $noir);
+				ImageLine($imgfin, 0, (int)($h + $hauteur_ligne[$ligne] + $padding / 2), $largeurfin - 1, (int)($h + $hauteur_ligne[$ligne] + $padding / 2), $noir);
 			}
 			$h+=$hauteur_ligne[$ligne] + $padding;
 		}
@@ -1466,7 +1466,7 @@ class expression_math extends expression
 		imagecopy($imgfin, $imgsup, $largeur - 6, 0, $largeursup - 6, 0, $largeursup, $hauteursup);
 		imagesetthickness($imgfin, 1);
 		imageline($imgfin, 0, 6, $largeur - 4, 6, $noir);
-		imagecopy($imgfin, $imgexp, ($largeur - $largeurexp) / 2, $hauteursup, 0, 0, $largeurexp, $hauteurexp);
+		imagecopy($imgfin, $imgexp, (int)(($largeur - $largeurexp) / 2), $hauteursup, 0, 0, $largeurexp, $hauteurexp);
 		$this->image = $imgfin;
 		$this->base_verticale = $baseexp + $hauteursup;
 	}
@@ -1539,8 +1539,8 @@ class expression_math extends expression
 		$blanc = imagecolorallocate($imgfin, 255, 255, 255);
 		$blanc = imagecolortransparent($imgfin, $blanc);
 		imagefilledrectangle($imgfin, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-		imagecopy($imgfin, $imgsup, ($largeur - $largeursup) / 2, 0, 0, 0, $largeursup, $hauteursup);
-		imagecopy($imgfin, $imgexp, ($largeur - $largeurexp) / 2, $hauteursup, 0, 0, $largeurexp, $hauteurexp);
+		imagecopy($imgfin, $imgsup, (int)(($largeur - $largeursup) / 2), 0, 0, 0, $largeursup, $hauteursup);
+		imagecopy($imgfin, $imgexp, (int)(($largeur - $largeurexp) / 2), $hauteursup, 0, 0, $largeurexp, $hauteurexp);
 		$this->image = $imgfin;
 		$this->base_verticale = $baseexp + $hauteursup;
 	}
@@ -1571,8 +1571,8 @@ class expression_math extends expression
 		$blanc = imagecolorallocate($imgfin, 255, 255, 255);
 		$blanc = imagecolortransparent($imgfin, $blanc);
 		imagefilledrectangle($imgfin, 0, 0, $largeur - 1, $hauteur - 1, $blanc);
-		imagecopy($imgfin, $imglim, ($largeur - $largeurlim) / 2, 0, 0, 0, $largeurlim, $hauteurlim);
-		imagecopy($imgfin, $imginf, ($largeur - $largeurinf) / 2, $hauteurlim, 0, 0, $largeurinf, $hauteurinf);
+		imagecopy($imgfin, $imglim, (int)(($largeur - $largeurlim) / 2), 0, 0, 0, $largeurlim, $hauteurlim);
+		imagecopy($imgfin, $imginf, (int)(($largeur - $largeurinf) / 2), $hauteurlim, 0, 0, $largeurinf, $hauteurinf);
 
 		$this->image = alignement2($imgfin, $baselim, $imgexp, $baseexp);
 		$this->base_verticale = max($baselim, $baseexp);
